Add filesystem cache for challenges
That should help if the server uses multiple processes instead of
threads (like when using passenger or uwsgi). It's not trivial to ensure
the same process gets both calls for a new issue. If the second call
hits another process it can't access the ThreadSafe::Cache because
that's just available in the same process.

Using files all processes on the same host can write new challenges and
delete it. Each challenge is written in a own file with the challenge id
as filename. That removes race conditions when multiple processes want
to access the challenges store. The first request would create a new
file and if it's verified the challenge file gets deleted. No other
process will write or modify the file so no additional locking is
required.

NOTE: when running multiple processes are used you might get race
conditions for writing to `store.yaml` if two domains are issued at the
same time or monitoring runs while a new domain is issued. It will
either result in outdated monitoring data or a failed domain issue.

It still don't help if you have a distributed server with multiple
hosts. In that cases using something like memcached or just a central
database would help. But that might be overkill for now.